Exhaustively proving the Collatz Conjecture in Haskell
I spent a bit of time playing with the Collatz conjecture. Got
interested in proving it for a whole range of values. Started out doing
it sequentially, and used that as a springboard to move into parallel
computation with Haskell. It works extremely well: Proved the
conjecture for all numbers up to 1 million in about 4 seconds.